+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2021-03-14
+
 ### Added
 - Show number of days for available transfers in status widget
 
@@ -111,7 +113,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial pre-release
 
-[Unreleased]: https://github.com/brotkrueml/typo3-jobrouter-process/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-jobrouter-process/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/brotkrueml/typo3-jobrouter-process/compare/v0.5.1...v1.0.0
 [0.5.1]: https://github.com/brotkrueml/typo3-jobrouter-process/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/brotkrueml/typo3-jobrouter-process/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/brotkrueml/typo3-jobrouter-process/compare/v0.4.0...v0.4.1